server/requirements.txt
bidict==0.21.4
click==8.0.3
Flask==2.0.2
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.12.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
paho-mqtt==1.6.1
python-engineio==4.3.0
python-socketio==5.4.1
schedule==1.1.0
simple-websocket==0.4.0
SQLAlchemy==1.4.26
Werkzeug==2.0.2
wsproto==1.0.0
